+++ src/test/java/org/stellar/sdk/responses/OperationDeserializerTest.java
@@ -46,6 +46,7 @@ public class OperationDeserializerTest extends TestCase {
     assertEquals(operation.getPagingToken(), "3936840037961729");
     assertEquals(operation.getId(), new Long(3936840037961729L));
     assertNull(operation.isTransactionSuccessful());
+    assertFalse(operation.getFunderMuxed().isPresent());
 
     assertEquals(operation.getAccount(), "GAR4DDXYNSN2CORG3XQFLAPWYKTUMLZYHYWV4Y2YJJ4JO6ZJFXMJD7PT");
     assertEquals(operation.getStartingBalance(), "299454.904954");
@@ -56,6 +57,56 @@ public class OperationDeserializerTest extends TestCase {
     assertEquals(operation.getLinks().getSelf().getHref(), "/operations/3936840037961729");
     assertEquals(operation.getLinks().getSucceeds().getHref(), "/operations?cursor=3936840037961729&order=desc");
     assertEquals(operation.getLinks().getTransaction().getHref(), "/transactions/75608563ae63757ffc0650d84d1d13c0f3cd4970a294a2a6b43e3f454e0f9e6d");
+  }
+
+  @Test
+  public void testDeserializeMuxedCreateAccountOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"/operations/3936840037961729/effects{?cursor,limit,order}\",\n" +
+        "      \"templated\": true\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"/operations?cursor=3936840037961729\\u0026order=asc\"\n" +
+        "    },\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"/operations/3936840037961729\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"/operations?cursor=3936840037961729\\u0026order=desc\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"/transactions/75608563ae63757ffc0650d84d1d13c0f3cd4970a294a2a6b43e3f454e0f9e6d\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"account\": \"GAR4DDXYNSN2CORG3XQFLAPWYKTUMLZYHYWV4Y2YJJ4JO6ZJFXMJD7PT\",\n" +
+        "  \"funder\": \"GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G\",\n" +
+        "  \"funder_muxed\": \"MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4\",\n" +
+        "  \"funder_muxed_id\": \"420\",\n" +
+        "  \"id\": 3936840037961729,\n" +
+        "  \"paging_token\": \"3936840037961729\",\n" +
+        "  \"source_account\": \"GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L\",\n" +
+        "  \"source_account_muxed\": \"MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA\",\n" +
+        "  \"source_account_muxed_id\": \"123\",\n" +
+        "  \"starting_balance\": \"299454.904954\",\n" +
+        "  \"type\": \"create_account\",\n" +
+        "  \"type_i\": 0\n" +
+        "}";
+    CreateAccountOperationResponse operation = (CreateAccountOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getAccount(), "GAR4DDXYNSN2CORG3XQFLAPWYKTUMLZYHYWV4Y2YJJ4JO6ZJFXMJD7PT");
+
+    assertEquals(operation.getSourceAccountMuxed().get(), new MuxedAccount(
+        "MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA",
+        "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L",
+        123l
+    ));
+
+    assertEquals(operation.getFunder(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getFunderMuxed().get().getAccountIdAddress(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getFunderMuxed().get().toString(), "MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4");
+    assertEquals(operation.getFunderMuxed().get().getId().longValue(), 420l);
   }
 
   @Test
@@ -187,6 +238,52 @@ public class OperationDeserializerTest extends TestCase {
     assertEquals(operation.isAuthorize(), true);
     assertEquals(operation.isAuthorizedToMaintainLiabilities(), false);
     assertEquals(operation.getAsset(), Asset.createNonNativeAsset("EUR", "GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM"));
+    assertFalse(operation.getTrusteeMuxed().isPresent());
+  }
+
+  @Test
+  public void testDeserializeMuxedAllowTrustOperation() {
+    String json = "{\n" +
+        "        \"_links\": {\n" +
+        "          \"self\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/operations/3602979345141761\"\n" +
+        "          },\n" +
+        "          \"transaction\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/transactions/1f265c075e8559ee4c21a32ae53337658e52d7504841adad4144c37143ea01a2\"\n" +
+        "          },\n" +
+        "          \"effects\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/operations/3602979345141761/effects\"\n" +
+        "          },\n" +
+        "          \"succeeds\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/effects?order=desc\\u0026cursor=3602979345141761\"\n" +
+        "          },\n" +
+        "          \"precedes\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/effects?order=asc\\u0026cursor=3602979345141761\"\n" +
+        "          }\n" +
+        "        },\n" +
+        "        \"id\": \"3602979345141761\",\n" +
+        "        \"paging_token\": \"3602979345141761\",\n" +
+        "        \"source_account\": \"GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM\",\n" +
+        "        \"type\": \"allow_trust\",\n" +
+        "        \"type_i\": 7,\n" +
+        "        \"asset_type\": \"credit_alphanum4\",\n" +
+        "        \"asset_code\": \"EUR\",\n" +
+        "        \"asset_issuer\": \"GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM\",\n" +
+        "        \"trustee\": \"GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G\",\n" +
+        "        \"trustee_muxed\": \"MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4\",\n" +
+        "        \"trustee_muxed_id\": \"420\",\n" +
+        "        \"trustor\": \"GDZ55LVXECRTW4G36EZPTHI4XIYS5JUC33TUS22UOETVFVOQ77JXWY4F\",\n" +
+        "        \"authorize\": true\n," +
+        "        \"authorize_to_maintain_liabilities\": false\n" +
+        "      }";
+
+    AllowTrustOperationResponse operation = (AllowTrustOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getTrustor(), "GDZ55LVXECRTW4G36EZPTHI4XIYS5JUC33TUS22UOETVFVOQ77JXWY4F");
+    assertEquals(operation.getTrustee(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getTrusteeMuxed().get().getAccountIdAddress(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getTrusteeMuxed().get().toString(), "MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4");
+    assertEquals(operation.getTrusteeMuxed().get().getId().longValue(), 420l);
   }
 
   @Test
@@ -271,6 +368,51 @@ public class OperationDeserializerTest extends TestCase {
     assertEquals(operation.getTrustor(), "GDZ55LVXECRTW4G36EZPTHI4XIYS5JUC33TUS22UOETVFVOQ77JXWY4F");
     assertEquals(operation.getLimit(), "922337203685.4775807");
     assertEquals(operation.getAsset(), Asset.createNonNativeAsset("EUR", "GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM"));
+    assertFalse(operation.getTrustorMuxed().isPresent());
+  }
+
+  @Test
+  public void testDeserializeMuxedChangeTrustOperation() {
+    String json = "{\n" +
+        "        \"_links\": {\n" +
+        "          \"self\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/operations/3602970755207169\"\n" +
+        "          },\n" +
+        "          \"transaction\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/transactions/8d409a788543895843d269c3f97a2d6a2ebca6e9f8f9a7ae593457b5c0ba6644\"\n" +
+        "          },\n" +
+        "          \"effects\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/operations/3602970755207169/effects\"\n" +
+        "          },\n" +
+        "          \"succeeds\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/effects?order=desc\\u0026cursor=3602970755207169\"\n" +
+        "          },\n" +
+        "          \"precedes\": {\n" +
+        "            \"href\": \"//horizon-testnet.stellar.org/effects?order=asc\\u0026cursor=3602970755207169\"\n" +
+        "          }\n" +
+        "        },\n" +
+        "        \"id\": \"3602970755207169\",\n" +
+        "        \"paging_token\": \"3602970755207169\",\n" +
+        "        \"source_account\": \"GDZ55LVXECRTW4G36EZPTHI4XIYS5JUC33TUS22UOETVFVOQ77JXWY4F\",\n" +
+        "        \"type\": \"change_trust\",\n" +
+        "        \"type_i\": 6,\n" +
+        "        \"asset_type\": \"credit_alphanum4\",\n" +
+        "        \"asset_code\": \"EUR\",\n" +
+        "        \"asset_issuer\": \"GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM\",\n" +
+        "        \"limit\": \"922337203685.4775807\",\n" +
+        "        \"trustee\": \"GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM\",\n" +
+        "        \"trustor\": \"GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G\",\n" +
+        "        \"trustor_muxed\": \"MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4\",\n" +
+        "        \"trustor_muxed_id\": \"420\"\n" +
+        "      }";
+
+    ChangeTrustOperationResponse operation = (ChangeTrustOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getTrustee(), "GDIROJW2YHMSFZJJ4R5XWWNUVND5I45YEWS5DSFKXCHMADZ5V374U2LM");
+    assertEquals(operation.getTrustor(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getTrustorMuxed().get().getAccountIdAddress(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getTrustorMuxed().get().toString(), "MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4");
+    assertEquals(operation.getTrustorMuxed().get().getId().longValue(), 420l);
   }
 
   @Test
@@ -374,36 +516,86 @@ public class OperationDeserializerTest extends TestCase {
   @Test
   public void testDeserializeAccountMergeOperation() {
     String json = "{\n" +
-            "        \"_links\": {\n" +
-            "          \"self\": {\n" +
-            "            \"href\": \"http://horizon-testnet.stellar.org/operations/3424497684189185\"\n" +
-            "          },\n" +
-            "          \"transaction\": {\n" +
-            "            \"href\": \"http://horizon-testnet.stellar.org/transactions/68f4d37780e2a85f5698b73977126a595dff99aed852f14a7eb39221ce1f96d5\"\n" +
-            "          },\n" +
-            "          \"effects\": {\n" +
-            "            \"href\": \"http://horizon-testnet.stellar.org/operations/3424497684189185/effects\"\n" +
-            "          },\n" +
-            "          \"succeeds\": {\n" +
-            "            \"href\": \"http://horizon-testnet.stellar.org/effects?order=desc\\u0026cursor=3424497684189185\"\n" +
-            "          },\n" +
-            "          \"precedes\": {\n" +
-            "            \"href\": \"http://horizon-testnet.stellar.org/effects?order=asc\\u0026cursor=3424497684189185\"\n" +
-            "          }\n" +
-            "        },\n" +
-            "        \"id\": \"3424497684189185\",\n" +
-            "        \"paging_token\": \"3424497684189185\",\n" +
-            "        \"source_account\": \"GD6GKRABNDVYDETEZJQEPS7IBQMERCN44R5RCI4LJNX6BMYQM2KPGGZ2\",\n" +
-            "        \"type\": \"account_merge\",\n" +
-            "        \"type_i\": 8,\n" +
-            "        \"account\": \"GD6GKRABNDVYDETEZJQEPS7IBQMERCN44R5RCI4LJNX6BMYQM2KPGGZ2\",\n" +
-            "        \"into\": \"GAZWSWPDQTBHFIPBY4FEDFW2J6E2LE7SZHJWGDZO6Q63W7DBSRICO2KN\"\n" +
-            "      }";
+        "        \"_links\": {\n" +
+        "          \"self\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/operations/3424497684189185\"\n" +
+        "          },\n" +
+        "          \"transaction\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/transactions/68f4d37780e2a85f5698b73977126a595dff99aed852f14a7eb39221ce1f96d5\"\n" +
+        "          },\n" +
+        "          \"effects\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/operations/3424497684189185/effects\"\n" +
+        "          },\n" +
+        "          \"succeeds\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/effects?order=desc\\u0026cursor=3424497684189185\"\n" +
+        "          },\n" +
+        "          \"precedes\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/effects?order=asc\\u0026cursor=3424497684189185\"\n" +
+        "          }\n" +
+        "        },\n" +
+        "        \"id\": \"3424497684189185\",\n" +
+        "        \"paging_token\": \"3424497684189185\",\n" +
+        "        \"source_account\": \"GD6GKRABNDVYDETEZJQEPS7IBQMERCN44R5RCI4LJNX6BMYQM2KPGGZ2\",\n" +
+        "        \"type\": \"account_merge\",\n" +
+        "        \"type_i\": 8,\n" +
+        "        \"account\": \"GD6GKRABNDVYDETEZJQEPS7IBQMERCN44R5RCI4LJNX6BMYQM2KPGGZ2\",\n" +
+        "        \"into\": \"GAZWSWPDQTBHFIPBY4FEDFW2J6E2LE7SZHJWGDZO6Q63W7DBSRICO2KN\"\n" +
+        "      }";
 
     AccountMergeOperationResponse operation = (AccountMergeOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
 
     assertEquals(operation.getAccount(), "GD6GKRABNDVYDETEZJQEPS7IBQMERCN44R5RCI4LJNX6BMYQM2KPGGZ2");
     assertEquals(operation.getInto(), "GAZWSWPDQTBHFIPBY4FEDFW2J6E2LE7SZHJWGDZO6Q63W7DBSRICO2KN");
+    assertFalse(operation.getIntoMuxed().isPresent());
+    assertFalse(operation.getAccountMuxed().isPresent());
+  }
+
+  @Test
+  public void testDeserializeMuxedAccountMergeOperation() {
+    String json = "{\n" +
+        "        \"_links\": {\n" +
+        "          \"self\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/operations/3424497684189185\"\n" +
+        "          },\n" +
+        "          \"transaction\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/transactions/68f4d37780e2a85f5698b73977126a595dff99aed852f14a7eb39221ce1f96d5\"\n" +
+        "          },\n" +
+        "          \"effects\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/operations/3424497684189185/effects\"\n" +
+        "          },\n" +
+        "          \"succeeds\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/effects?order=desc\\u0026cursor=3424497684189185\"\n" +
+        "          },\n" +
+        "          \"precedes\": {\n" +
+        "            \"href\": \"http://horizon-testnet.stellar.org/effects?order=asc\\u0026cursor=3424497684189185\"\n" +
+        "          }\n" +
+        "        },\n" +
+        "        \"id\": \"3424497684189185\",\n" +
+        "        \"paging_token\": \"3424497684189185\",\n" +
+        "        \"source_account\": \"GD6GKRABNDVYDETEZJQEPS7IBQMERCN44R5RCI4LJNX6BMYQM2KPGGZ2\",\n" +
+        "        \"type\": \"account_merge\",\n" +
+        "        \"type_i\": 8,\n" +
+        "        \"account\": \"GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L\",\n" +
+        "        \"account_muxed\": \"MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA\",\n" +
+        "        \"account_muxed_id\": \"123\",\n" +
+        "        \"into\": \"GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G\",\n" +
+        "        \"into_muxed\": \"MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4\",\n" +
+        "        \"into_muxed_id\": \"420\"\n" +
+        "      }";
+
+    AccountMergeOperationResponse operation = (AccountMergeOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getAccount(), "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L");
+    assertEquals(operation.getInto(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+
+    assertEquals(operation.getAccountMuxed().get(), new MuxedAccount(
+        "MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA",
+        "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L",
+        123l
+    ));
+    assertEquals(operation.getIntoMuxed().get().getAccountIdAddress(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getIntoMuxed().get().toString(), "MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4");
+    assertEquals(operation.getIntoMuxed().get().getId().longValue(), 420l);
   }
 
   @Test
@@ -543,8 +735,11 @@ public class OperationDeserializerTest extends TestCase {
 
     PathPaymentStrictSendOperationResponse operation = (PathPaymentStrictSendOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
 
+    assertFalse(operation.getSourceAccountMuxed().isPresent());
     assertEquals(operation.getFrom(), "GC45JH537XZD4DY4WTV5PCUJL4KPOIE4WMGX5OP5KSPS2OLGRUOVVIGD");
+    assertFalse(operation.getFromMuxed().isPresent());
     assertEquals(operation.getTo(), "GC45JH537XZD4DY4WTV5PCUJL4KPOIE4WMGX5OP5KSPS2OLGRUOVVIGD");
+    assertFalse(operation.getToMuxed().isPresent());
     assertEquals(operation.getAmount(), "2.5000000");
     assertEquals(operation.getSourceAmount(), "1.1777000");
     assertEquals(operation.getDestinationMin(), "1.1779523");
@@ -555,6 +750,64 @@ public class OperationDeserializerTest extends TestCase {
             Asset.createNonNativeAsset("CNYMNL", "GAREELUB43IRHWEASCFBLKHURCGMHE5IF6XSE7EXDLACYHGRHM43RFOX")
     ));
     assertEquals(operation.getSourceAsset(), Asset.createNonNativeAsset("XRP", "GBVOL67TMUQBGL4TZYNMY3ZQ5WGQYFPFD5VJRWXR72VA33VFNL225PL5"));
+  }
+
+  @Test
+  public void testDeserializeMuxedPathPaymentStrictSendOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"https://horizon.stellar.org/operations/75252830662840321\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"https://horizon.stellar.org/transactions/fb2f5655c70a459220ac09eb3d6870422b58dcf5c5ffb5e5b21817b4d248826e\"\n" +
+        "    },\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"https://horizon.stellar.org/operations/75252830662840321/effects\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"https://horizon.stellar.org/effects?order=desc\\u0026cursor=75252830662840321\"\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"https://horizon.stellar.org/effects?order=asc\\u0026cursor=75252830662840321\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"id\": \"75252830662840321\",\n" +
+        "  \"paging_token\": \"75252830662840321\",\n" +
+        "  \"source_account\": \"GC45JH537XZD4DY4WTV5PCUJL4KPOIE4WMGX5OP5KSPS2OLGRUOVVIGD\",\n" +
+        "  \"type\": \"path_payment_strict_send\",\n" +
+        "  \"type_i\": 13,\n" +
+        "  \"created_at\": \"2018-04-24T12:58:12Z\",\n" +
+        "  \"transaction_hash\": \"fb2f5655c70a459220ac09eb3d6870422b58dcf5c5ffb5e5b21817b4d248826e\",\n" +
+        "  \"asset_type\": \"native\",\n" +
+        "  \"from\": \"GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L\",\n" +
+        "  \"from_muxed\": \"MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA\",\n" +
+        "  \"from_muxed_id\": \"123\",\n" +
+        "  \"to\": \"GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G\",\n" +
+        "  \"to_muxed\": \"MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4\",\n" +
+        "  \"to_muxed_id\": \"420\",\n" +
+        "  \"amount\": \"2.5000000\",\n" +
+        "  \"path\": [{\"asset_type\": \"native\"}, {\"asset_type\": \"credit_alphanum4\", \"asset_code\": \"CNY\", \"asset_issuer\": \"GAREELUB43IRHWEASCFBLKHURCGMHE5IF6XSE7EXDLACYHGRHM43RFOX\"}, {\"asset_type\": \"credit_alphanum12\", \"asset_code\": \"CNYMNL\", \"asset_issuer\": \"GAREELUB43IRHWEASCFBLKHURCGMHE5IF6XSE7EXDLACYHGRHM43RFOX\"}],\n" +
+        "  \"source_amount\": \"1.1777000\",\n"+
+        "  \"destination_min\": \"1.1779523\",\n" +
+        "  \"source_asset_type\": \"credit_alphanum4\",\n" +
+        "  \"source_asset_code\": \"XRP\",\n" +
+        "  \"source_asset_issuer\": \"GBVOL67TMUQBGL4TZYNMY3ZQ5WGQYFPFD5VJRWXR72VA33VFNL225PL5\"\n" +
+        "}";
+
+    PathPaymentStrictSendOperationResponse operation = (PathPaymentStrictSendOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getFrom(), "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L");
+    assertEquals(operation.getFromMuxed().get(), new MuxedAccount(
+        "MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA",
+        "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L",
+        123l
+    ));
+
+    assertEquals(operation.getTo(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getToMuxed().get().getAccountIdAddress(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getToMuxed().get().toString(), "MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4");
+    assertEquals(operation.getToMuxed().get().getId().longValue(), 420l);
   }
 
   @Test
@@ -743,6 +996,164 @@ public class OperationDeserializerTest extends TestCase {
   }
 
   @Test
+  public void testDeserializeEndSponsoringFutureReservesOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"/operations/12884914177/effects/{?cursor,limit,order}\",\n" +
+        "      \"templated\": true\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=asc\"\n" +
+        "    },\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"/operations/12884914177\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=desc\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"/transactions/12884914176\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"id\": 12884914177,\n" +
+        "  \"paging_token\": \"12884914177\",\n" +
+        "  \"type_i\": 17,\n" +
+        "  \"type\": \"end_sponsoring_future_reserves\",\n" +
+        "  \"begin_sponsor\": \"GDRW375MAYR46ODGF2WGANQC2RRZL7O246DYHHCGWTV2RE7IHE2QUQLD\"\n" +
+        "}";
+
+    EndSponsoringFutureReservesOperationResponse operation = (EndSponsoringFutureReservesOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getId(), new Long(12884914177L));
+    assertEquals(operation.getBeginSponsor(), "GDRW375MAYR46ODGF2WGANQC2RRZL7O246DYHHCGWTV2RE7IHE2QUQLD");
+    assertFalse(operation.getBeginSponsorMuxed().isPresent());
+    assertEquals(operation.getType(), "end_sponsoring_future_reserves");
+  }
+
+  @Test
+  public void testDeserializeMuxedEndSponsoringFutureReservesOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"/operations/12884914177/effects/{?cursor,limit,order}\",\n" +
+        "      \"templated\": true\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=asc\"\n" +
+        "    },\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"/operations/12884914177\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=desc\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"/transactions/12884914176\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"id\": 12884914177,\n" +
+        "  \"paging_token\": \"12884914177\",\n" +
+        "  \"type_i\": 17,\n" +
+        "  \"type\": \"end_sponsoring_future_reserves\",\n" +
+        "  \"begin_sponsor\": \"GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G\",\n" +
+        "  \"begin_sponsor_muxed\": \"MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4\",\n" +
+        "  \"begin_sponsor_muxed_id\": \"420\"\n" +
+        "}";
+
+    EndSponsoringFutureReservesOperationResponse operation = (EndSponsoringFutureReservesOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getId(), new Long(12884914177L));
+    assertEquals(operation.getBeginSponsor(), "GAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472ZM7G");
+    assertEquals(operation.getBeginSponsorMuxed().get().toString(), "MAVH5JM5OKXGMQDS7YPRJ4MQCPXJUGH26LYQPQJ4SOMOJ4SXY472YAAAAAAAAAABUSON4");
+    assertEquals(operation.getBeginSponsorMuxed().get().getId().longValue(), 420l);
+    assertEquals(operation.getType(), "end_sponsoring_future_reserves");
+  }
+
+  @Test
+  public void testDeserializeClaimClaimableBalanceOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"/operations/12884914177/effects/{?cursor,limit,order}\",\n" +
+        "      \"templated\": true\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=asc\"\n" +
+        "    },\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"/operations/12884914177\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=desc\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"/transactions/12884914176\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"id\": 12884914177,\n" +
+        "  \"paging_token\": \"12884914177\",\n" +
+        "  \"type_i\": 15,\n" +
+        "  \"type\": \"claim_claimable_balance\",\n" +
+        "  \"claimant\": \"GDRW375MAYR46ODGF2WGANQC2RRZL7O246DYHHCGWTV2RE7IHE2QUQLD\",\n" +
+        "  \"balance_id\": \"00000000178826fbfe339e1f5c53417c6fedfe2c05e8bec14303143ec46b38981b09c3f9\"\n" +
+        "}";
+
+    ClaimClaimableBalanceOperationResponse operation = (ClaimClaimableBalanceOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getId(), new Long(12884914177L));
+    assertEquals(operation.getBalanceId(), "00000000178826fbfe339e1f5c53417c6fedfe2c05e8bec14303143ec46b38981b09c3f9");
+    assertEquals(operation.getClaimant(), "GDRW375MAYR46ODGF2WGANQC2RRZL7O246DYHHCGWTV2RE7IHE2QUQLD");
+    assertFalse(operation.getClaimantMuxed().isPresent());
+    assertEquals(operation.getType(), "claim_claimable_balance");
+  }
+
+  @Test
+  public void testDeserializeMuxedClaimClaimableBalanceOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"/operations/12884914177/effects/{?cursor,limit,order}\",\n" +
+        "      \"templated\": true\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=asc\"\n" +
+        "    },\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"/operations/12884914177\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=desc\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"/transactions/12884914176\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"id\": 12884914177,\n" +
+        "  \"paging_token\": \"12884914177\",\n" +
+        "  \"type_i\": 15,\n" +
+        "  \"type\": \"claim_claimable_balance\",\n" +
+        "  \"claimant\": \"GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L\",\n" +
+        "  \"claimant_muxed\": \"MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA\",\n" +
+        "  \"claimant_muxed_id\": \"123\",\n" +
+        "  \"balance_id\": \"00000000178826fbfe339e1f5c53417c6fedfe2c05e8bec14303143ec46b38981b09c3f9\"\n" +
+        "}";
+
+    ClaimClaimableBalanceOperationResponse operation = (ClaimClaimableBalanceOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getId(), new Long(12884914177L));
+    assertEquals(operation.getBalanceId(), "00000000178826fbfe339e1f5c53417c6fedfe2c05e8bec14303143ec46b38981b09c3f9");
+    assertEquals(operation.getClaimant(), "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L");
+    assertEquals(operation.getClaimantMuxed().get(), new MuxedAccount(
+        "MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA",
+        "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L",
+        123l
+    ));
+    assertEquals(operation.getType(), "claim_claimable_balance");
+  }
+
+
+  @Test
   public void testDeserializeClawbackClaimableBalanceOperation() {
     String json = "{\n" +
         "  \"_links\": {\n" +
@@ -817,6 +1228,50 @@ public class OperationDeserializerTest extends TestCase {
     assertEquals(operation.getAssetIssuer(), "GCWVFBJ24754I5GXG4JOEB72GJCL3MKWC7VAEYWKGQHPVH3ENPNBSKWS");
     assertEquals(operation.getAssetCode(), "EUR");
     assertEquals(operation.getAssetType(), "credit_alphanum4");
+    assertFalse(operation.getFromMuxed().isPresent());
+  }
+
+  @Test
+  public void testDeserializeMuxedClawbackOperation() {
+    String json = "{\n" +
+        "  \"_links\": {\n" +
+        "    \"effects\": {\n" +
+        "      \"href\": \"/operations/12884914177/effects/{?cursor,limit,order}\",\n" +
+        "      \"templated\": true\n" +
+        "    },\n" +
+        "    \"precedes\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=asc\"\n" +
+        "    },\n" +
+        "    \"self\": {\n" +
+        "      \"href\": \"/operations/12884914177\"\n" +
+        "    },\n" +
+        "    \"succeeds\": {\n" +
+        "      \"href\": \"/operations?cursor=12884914177\\u0026order=desc\"\n" +
+        "    },\n" +
+        "    \"transaction\": {\n" +
+        "      \"href\": \"/transactions/12884914176\"\n" +
+        "    }\n" +
+        "  },\n" +
+        "  \"id\": 12884914177,\n" +
+        "  \"paging_token\": \"12884914177\",\n" +
+        "  \"type_i\": 19,\n" +
+        "  \"type\": \"clawback\",\n" +
+        "  \"asset_type\": \"credit_alphanum4\",\n" +
+        "  \"asset_code\": \"EUR\",\n" +
+        "  \"asset_issuer\": \"GCWVFBJ24754I5GXG4JOEB72GJCL3MKWC7VAEYWKGQHPVH3ENPNBSKWS\",\n" +
+        "  \"from\": \"GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L\",\n" +
+        "  \"from_muxed\": \"MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA\",\n" +
+        "  \"from_muxed_id\": \"123\"\n" +
+        "}";
+
+    ClawbackOperationResponse operation = (ClawbackOperationResponse) GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
+
+    assertEquals(operation.getFrom(), "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L");
+    assertEquals(operation.getFromMuxed().get(), new MuxedAccount(
+        "MBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA6AAAAAAAAAAAPN7BA",
+        "GBB4JST32UWKOLGYYSCEYBHBCOFL2TGBHDVOMZP462ET4ZRD4ULA7S2L",
+        123l
+    ));
   }
 
   @Test